openapi: relocate unit tests
Separate unit tests in different directories according to OpenAPI
version.

Extracted from #2152.
