+++ addOns/openapi/src/test/java/org/zaproxy/zap/extension/openapi/v2/OpenApiPrimitivesInBodyUnitTest.java
@@ -3,7 +3,7 @@
  *
  * ZAP is an HTTP/HTTPS proxy for assessing web application security.
  *
- * Copyright 2017 The ZAP Development Team
+ * Copyright 2019 The ZAP Development Team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,20 +17,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.zap.extension.openapi;
+package org.zaproxy.zap.extension.openapi.v2;
 
 import static fi.iki.elonen.NanoHTTPD.newFixedLengthResponse;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 
-import fi.iki.elonen.NanoHTTPD.IHTTPSession;
-import fi.iki.elonen.NanoHTTPD.Response;
+import fi.iki.elonen.NanoHTTPD;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.Test;
 import org.parosproxy.paros.network.HttpMessage;
 import org.parosproxy.paros.network.HttpSender;
+import org.zaproxy.zap.extension.openapi.AbstractOpenApiTest;
 import org.zaproxy.zap.extension.openapi.converter.Converter;
 import org.zaproxy.zap.extension.openapi.converter.swagger.SwaggerConverter;
 import org.zaproxy.zap.extension.openapi.converter.swagger.SwaggerException;
@@ -38,22 +37,23 @@ import org.zaproxy.zap.extension.openapi.network.RequesterListener;
 import org.zaproxy.zap.extension.openapi.network.Requestor;
 import org.zaproxy.zap.testutils.NanoServerHandler;
 
-public class OpenApiPre2dot0UnitTest extends AbstractOpenApiTest {
+public class OpenApiPrimitivesInBodyUnitTest extends AbstractOpenApiTest {
+
     @Test
-    public void shouldExplorePetStore1_2()
+    public void shouldExploreBodiesWithPrimitiveValues()
             throws NullPointerException, IOException, SwaggerException {
-        String test = "/PetStore_1_2_defn/";
+        String test = "/OpenApi_defn_body_with_primitives/";
 
         this.nano.addHandler(
                 new NanoServerHandler(test) {
                     @Override
-                    protected Response serve(IHTTPSession session) {
+                    protected NanoHTTPD.Response serve(NanoHTTPD.IHTTPSession session) {
                         String response;
                         String uri = session.getUri();
                         if (uri.endsWith("defn.json")) {
                             response =
                                     getHtml(
-                                            "PetStore_1_2_defn.json",
+                                            "OpenApi_defn_body_with_primitives.json",
                                             new String[][] {
                                                 {"PORT", String.valueOf(nano.getListeningPort())}
                                             });
@@ -85,17 +85,28 @@ public class OpenApiPre2dot0UnitTest extends AbstractOpenApiTest {
         requestor.addListener(listener);
         requestor.run(converter.getRequestModels());
 
-        checkPetStore2dot0Requests(accessedUrls, "localhost:" + nano.getListeningPort());
+        checkRequests(accessedUrls, "localhost:" + nano.getListeningPort());
     }
 
-    private void checkPetStore2dot0Requests(Map<String, String> accessedUrls, String host) {
+    private void checkRequests(Map<String, String> accessedUrls, String host) {
         // Check all of the expected URLs have been accessed and with the right data
-        assertTrue(accessedUrls.containsKey("POST http://" + host + "/store/order"));
+        assertEquals("true", accessedUrls.get("POST http://" + host + "/api/boolean"));
+        assertEquals("[true,true]", accessedUrls.get("POST http://" + host + "/api/booleans"));
+
+        assertEquals("1.2", accessedUrls.get("POST http://" + host + "/api/double"));
+        assertEquals("[1.2,1.2]", accessedUrls.get("POST http://" + host + "/api/doubles"));
+
+        assertEquals("10", accessedUrls.get("POST http://" + host + "/api/integer"));
+        assertEquals("[10,10]", accessedUrls.get("POST http://" + host + "/api/integers"));
+
+        assertEquals("10", accessedUrls.get("POST http://" + host + "/api/long"));
+        assertEquals("[10,10]", accessedUrls.get("POST http://" + host + "/api/longs"));
+
+        assertEquals("\"John Doe\"", accessedUrls.get("POST http://" + host + "/api/string"));
         assertEquals(
-                "{\"id\":10,\"petId\":10,\"quantity\":10,\"status\":\"placed\",\"shipDate\":\"1970-01-01T00:00:00.001Z\"}",
-                accessedUrls.get("POST http://" + host + "/store/order"));
-        assertTrue(accessedUrls.containsKey("GET http://" + host + "/store/order/orderId"));
-        assertTrue(accessedUrls.containsKey("DELETE http://" + host + "/store/order/orderId"));
-        assertEquals(3, accessedUrls.size());
+                "[\"John Doe\",\"John Doe\"]",
+                accessedUrls.get("POST http://" + host + "/api/strings"));
+
+        assertEquals(10, accessedUrls.size());
     }
 }
