+++ addOns/openapi/src/test/resources/org/zaproxy/zap/extension/openapi/v1/PetStore_defn.json
