+++ addOns/openapi/src/test/resources/org/zaproxy/zap/extension/openapi/v2/OpenApi_defn_body_with_primitives.json
