+++ addOns/openapi/src/test/resources/org/zaproxy/zap/extension/openapi/v2/PetStore_defn_no_host.json
